requirements.txt
Django==5.2.5
asgiref==3.11.0
sqlparse==0.5.5
python-dotenv==1.1.1
whitenoise==6.11.0
gunicorn==23.0.0
openai==1.99.9
requests==2.32.4
slack_sdk==3.36.0
django-encrypted-model-fields==0.6.5
django-picklefield==3.4.0
django-q2==1.8.0
psycopg2-binary==2.9.11
